Account for distance being 0 if exact lat lng match
if league id is 0.00 mi away they were being filtered out of the list
because 0 is falsy. We have to explicitly check if the id is not in the
mapping. This happens if lat,lng (1,5) is saved and if you go to search
and look for exactly (1,5). That will give a distance of 0.00